Add useApiRequest react hook to make api requests in components
Sometimes we don’t need to share data across React components and don’t need to store data in Redux.
For these times, we can use a hook that just keeps the data private to that component.

This is just a simple wrapper around `useState`/`useEffect` that makes the API request using `apiRequest()`
It takes an object in the shape of the method signature for `apiRequest` and a list of dependencies
that `useEffect` will trigger on changes of.

Finally, there was some restructuring of the directory structure to allow for an additional function
from another file to be exported properly as the npm package