API-1768 Add prettier to project
Prettier helps with consistent code formatting. We want to enforce some
code formatting rules in our latest ApiUtils package.

The new NPM package Prettier is added to this project and we also make
the eslint work together with prettier. Therefore, the Gitlab CI
pipeline will warn us about the violations.

To format the code, we can run
`npm run prettierFormat`
To only check the violations, we can run
`npx prettier --check .`

The .prettierrc.json is based on the manage repo. We want to ignore the
non-JS files in the formatting, so the .prettierignore is added.

Related violations are corrected in this commit.