Fix bug with action batch responses returning errors
I noticed that creating an action batch that returns an error that the
library continues to make API requests to the back-end in a loop, with
no throttling. This is because the API response that is being checked is
the initial POST response, not the current GET status response. If the
initial response doesn't return an error, then the error condition never
gets tripped.

This commit fixes that bug, and adds a test. I also bumped the version
of the package by a patch-level change.